+++ cpuMergesort.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 #include <time.h>
 
-extern int mergesort_cpu(int *list, int *sorted, int n);
+extern int mergesort_cpu(double *list, double *sorted, int n);
